\name{Initialize}
\alias{Initialize}
\title{
Initialize internal variables in a WrapData object
}
\description{
Internal function used by other functions in the package. Used to initialize parameters 
in the wrapped data object.
}
\usage{
Initialize(inbox)
}
\arguments{
  \item{inbox}{ Object returned by the WrapData function.}
}
\details{
   This is an internal function used ony by the other functions of this package. 
}
\value{
   An R object containing information on the data to be transformed.
}
\author{
Tridivesh Jena, Zementis, Inc. 
}

\seealso{
  \code{\link{WrapData}}
}

